Moves resource api methods out of client file.
